Created wizardHelperProcessModal mixin to reduce code duplication
